Progreso en la serializacion de autores

+++ devocionales-spring/src/main/java/dylan/devocionalesspring/repositorios/ComentarioRepositorio.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ComentarioRepositorio extends JpaRepository<Comentario, Long> {
-    List<Comentario> findByDevocionalId(int devocionalId);
+    List<Comentario> findByDevocionalId(Long devocionalId);
 }
